fix: Fall back to local data when fetch api failed
Fetch data from API may fail in both production and
local development environements.
In local env, the fetch fails due to CORS restriction of the browser.
In production, the netlify proxy put timeout restriction
on the proxy request for 10s, and it may fail to fetch data
before the timeout.
Therefore, when the data could not be fetched,
use the local copy one to ensure that the website works correctly.